vocab-backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
pydantic==1.10.7
python-dotenv==1.0.0
jsonschema==4.17.3

streamlit>=1.30.0
requests>=2.31.0
pillow>=10.1.0


boto3>=1.34.0
chromadb>=0.4.22
boto3